исправление замечаний CodeFactor
