+++ Zilon.Core/Zilon.Core.Tests.Common/Schemes/TestSectorTransitionSubScheme.cs
@@ -1,0 +1,9 @@
+﻿using Zilon.Core.Schemes;
+
+namespace Zilon.Core.Tests.Common.Schemes
+{
+    public class TestSectorTransitionSubScheme : ISectorTransitionSubScheme
+    {
+        public string SectorLevelSid { get; set; }
+    }
+}
