+++ Zilon.Core/Zilon.Core/Schemes/SectorTransitionSubScheme.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Zilon.Core.Schemes
+{
+    public sealed class SectorTransitionSubScheme : ISectorTransitionSubScheme
+    {
+        [JsonProperty]
+        public string SectorLevelSid { get; private set; }
+    }
+}
